user-service/pyproject.toml
[tool.poetry]
name = "user-service"
version = "0.1.0"
description = ""
authors = ["eedu7 <mueedahmad067@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.115.5"}
uvicorn = "^0.32.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
alembic = "^1.14.0"
pydantic = "^2.9.2"
pydantic-settings = "^2.6.1"
pytest = "^8.3.3"
pyjwt = "^2.9.0"
redis = "^5.2.0"
email-validator = "^2.2.0"
passlib = "^1.7.4"
bcrypt = "^4.2.0"
faker = "^33.0.0"
pytest-asyncio = "^0.24.0"
asyncpg = "^0.30.0"
python-jose = "^3.3.0"
contextvars = "^2.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
